Add bundle size validations + cleanup community entry points
