+++ package.json
@@ -39,6 +39,9 @@
     "@nx/js": "18.3.4",
     "@nx/vite": "18.3.4",
     "@nx/workspace": "18.3.4",
+    "@size-limit/esbuild": "^11.1.6",
+    "@size-limit/esbuild-why": "^11.1.6",
+    "@size-limit/file": "^11.1.6",
     "@swc-node/register": "1.8.0",
     "@swc/cli": "0.1.62",
     "@swc/core": "1.3.96",
@@ -71,8 +74,8 @@
     "esbuild-plugin-umd-wrapper": "^2.0.0",
     "eslint": "^9.14.0",
     "eslint-plugin-check-file": "^2.8.0",
-    "eslint-plugin-sonarjs": "^2.0.4",
     "eslint-plugin-react-hooks": "^5.1.0-rc-df783f9ea1-20240708",
+    "eslint-plugin-sonarjs": "^2.0.4",
     "glob": "8.0.3",
     "globals": "^15.12.0",
     "html-minifier-terser": "^7.2.0",
@@ -106,6 +109,7 @@
     "react-instantsearch-dom": "^6.40.1",
     "react-query": "^3.39.3",
     "rimraf": "3.0.2",
+    "size-limit": "^11.1.6",
     "sucrase": "^3.32.0",
     "timezone-mock": "^1.3.6",
     "ts-jest": "^29.1.0",
