+++ packages/ag-charts-community/src/main-modules.ts
@@ -1,23 +1,1 @@
-import type { IntegratedModule } from 'ag-charts-types';
-
-import { AgCharts } from './api/agCharts';
-import { registerInbuiltModules } from './chart/factory/registerInbuiltModules';
-import * as _Scene from './integrated-charts-scene';
-import * as _Theme from './integrated-charts-theme';
-import * as _Util from './integrated-charts-util';
-import { VERSION } from './version';
-
-export { AgCharts } from './api/agCharts';
-export { VERSION } from './version';
-export { registerInbuiltModules as setupCommunityModules } from './chart/factory/registerInbuiltModules';
-
-export const AgChartsCommunityModule = {
-    VERSION,
-    _Scene,
-    _Theme,
-    _Util,
-    create: AgCharts.create.bind(AgCharts),
-    createSparkline: AgCharts.__createSparkline.bind(AgCharts),
-    setup: registerInbuiltModules,
-    isEnterprise: false,
-} satisfies IntegratedModule;
+// Entry point to implement and test our tree-shaking abilities
