+++ packages/ag-charts-enterprise/package.json
@@ -7,13 +7,12 @@
   "module": "./dist/package/main.esm.mjs",
   "exports": {
     ".": {
-      "import": "./dist/package/main.esm.mjs",
-      "require": "./dist/package/main.cjs.js",
       "types": "./dist/types/src/main.d.ts",
+      "require": "./dist/package/main.cjs.js",
+      "import": "./dist/package/main.esm.mjs",
       "default": "./dist/package/main.cjs.js"
     }
   },
-  "private": false,
   "scripts": {
     "lint:prune": "npx ts-prune -p tsconfig.lib.json"
   },
